Separate `dhcp` and `dhcpd` bin commands into `dhcp-cli.js` and `dhcpd-cli.js`
